Python: type objects, runtime defaults, bigint literals (353→369)
Every host exception that was escaping to scripts is now gone; the remaining
failures are all Python-level.

First-class type objects
- `int`, `str`, `list`, … were constructor functions, which makes both
  `type(x) is int` and `isinstance(x, int)` false. They are now `PyType`
  singletons that are callable, comparable by identity, and know their own
  instance test. `type(x)` returns the type object — the user's class for an
  instance, the exception class for an exception. `bool` is treated as a
  subtype of `int`, as Python defines it.

Default values
- Defaults were folded at compile time, which rejected every non-literal
  default and would have given `def f(x=[])` fresh lists per call. They are now
  evaluated once at definition time in the enclosing scope and captured on the
  function object, so a mutable default is shared across calls as Python
  specifies.

Correctness
- Integer literals parse as BigInteger throughout; `long.Parse` in the tokenizer
  made every large literal a host OverflowException
- A captured local is now read through its cell: a nested function's write went
  to the cell while the defining scope kept reading its stale slot
- Decorator expressions evaluate top to bottom while the decorators apply bottom
  up; both were happening in reverse
- `0 ** -1` raises ZeroDivisionError instead of returning infinity
- `d.update(d)` snapshots before writing back
- Built-in methods enforce an upper argument bound, not just a lower one
- `bytes(-1)`, `len(huge_range)`, `itertools.count(1.5)` and a malformed
  `dict()` sequence all raise Python errors rather than host exceptions
- Lambda parameters cannot be annotated, so `:` ends the list

369 of 558 fixtures passing. Shell suites unchanged and green.

+++ src/Monty/Parsing/Parser.cs
@@ -1,4 +1,5 @@
 using System.Globalization;
+using System.Numerics;
 
 namespace Monty.Parsing;
 
@@ -1171,7 +1172,7 @@ public sealed class Parser
         {
             case TokenKind.Integer:
                 _index++;
-                return At(new Literal(long.Parse(token.Text, CultureInfo.InvariantCulture)), token);
+                return At(new Literal(BigInteger.Parse(token.Text, CultureInfo.InvariantCulture)), token);
 
             case TokenKind.Float:
                 _index++;
